Remove unnecessary benchmark runs The benchmarks were running twice in master for every YAML config and the 1st result was discarded.
This commit changes the way mathics-bench works. Before of it,
calling mathics-bench with a ref argument would run the benchmarks
in master and in ref, now it only runs in ref.
(ref defaults to "master")

I removed a comment from @mmatera which was saying that we don't
checkout to master at the end of the benchmark so it not recompile
mathics-core when running multiple benchmarks at the same branch
because now it's true by default as the checkout to ref is above
the setup (install, compile, ...).